requirements.txt
pymongo
python-dotenv

scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
joblib==1.4.2
threadpoolctl==3.5.0
imbalanced-learn==0.12.0